:sparkles: Feat: AI Service 디테일 페이지 서비스 소개 부분 컴포넌트화
